docker/requirements.txt
fastapi
uvicorn[standard]
pydantic

langchain-core
langgraph

langchain-community
langchain-huggingface

qdrant-client

pypdf
unstructured

torch
transformers
sentence-transformers

python-dotenv

langfuse